COM-3685: rename context
